+++ packages/mix/lib/experimental.dart
@@ -1,0 +1,16 @@
+///   /\\\\            /\\\\  /\\\\\\\\\\\  /\\\       /\\\
+///   \/\\\\\\        /\\\\\\ \/////\\\///  \///\\\   /\\\/
+///    \/\\\//\\\    /\\\//\\\     \/\\\       \///\\\\\\/
+///     \/\\\\///\\\/\\\/ \/\\\     \/\\\         \//\\\\
+///      \/\\\  \///\\\/   \/\\\     \/\\\          \/\\\\
+///       \/\\\    \///     \/\\\     \/\\\          /\\\\\\
+///        \/\\\             \/\\\     \/\\\        /\\\////\\\
+///         \/\\\             \/\\\  /\\\\\\\\\\\  /\\\/   \///\\\
+///          \///              \///  \///////////  \///       \///
+///
+///                        https://fluttermix.com
+///
+
+library mix_experimental;
+
+export 'src/core/internal/experimental/mix_builder.dart';
