+++ packages/mix/lib/src/core/styled_widget.dart
@@ -1,12 +1,10 @@
 import 'package:flutter/widgets.dart';
 
-import '../modifiers/internal/render_widget_modifier.dart';
 import '../variants/widget_state_variant.dart';
 import '../widgets/pressable_widget.dart';
 import 'factory/mix_data.dart';
-import 'factory/mix_provider.dart';
 import 'factory/style_mix.dart';
-import 'modifier.dart';
+import 'internal/experimental/mix_builder.dart';
 import 'widget_state/widget_state_controller.dart';
 
 /// An abstract widget for applying custom styles.
@@ -93,7 +91,7 @@ class SpecBuilder extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     Widget current = Builder(
-      builder: (_) => MixComposer(
+      builder: (_) => MixBuilder(
         inherit: inherit,
         style: style,
         orderOfModifiers: orderOfModifiers,
@@ -110,60 +108,5 @@ class SpecBuilder extends StatelessWidget {
 
     // Otherwise, directly build the mixed child widget
     return current;
-  }
-}
-
-class MixComposer extends StatelessWidget {
-  const MixComposer({
-    super.key,
-    this.inherit = false,
-    required this.style,
-    this.orderOfModifiers = const [],
-    required this.builder,
-  });
-
-  final bool inherit;
-  final Style style;
-  final List<Type> orderOfModifiers;
-  final Widget Function(BuildContext) builder;
-
-  Widget _applyModifiers(MixData mix, Widget child) {
-    // Get the list of WidgetModifierAttribute from the mix
-    final modifiers = mix
-        .whereType<WidgetModifierSpecAttribute>()
-        .map((e) => e.resolve(mix))
-        .toList();
-
-    // If the mix is animated, use RenderAnimatedModifiers, otherwise use RenderModifiers
-    return mix.isAnimated
-        ? RenderAnimatedModifiers(
-            modifiers: modifiers,
-            duration: mix.animation!.duration,
-            mix: mix,
-            orderOfModifiers: orderOfModifiers,
-            curve: mix.animation!.curve,
-            child: child,
-          )
-        : RenderModifiers(
-            modifiers: modifiers,
-            mix: mix,
-            orderOfModifiers: orderOfModifiers,
-            child: child,
-          );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final inheritedMix = inherit ? Mix.maybeOfInherited(context) : null;
-    // Get the mix from the style
-    final mix = style.of(context);
-    // Merge the inherited mix with the current mix, or use the current mix if no inherited mix
-    final mergedMix = inheritedMix?.merge(mix) ?? mix;
-
-    // Return a Mix widget with the merged mix and the child widget with modifiers applied
-    return Mix(
-      data: mergedMix,
-      child: _applyModifiers(mergedMix, Builder(builder: builder)),
-    );
   }
 }
